+++ queries/processed/elixir/injections.scm
@@ -62,24 +62,8 @@
 ((sigil
   (sigil_name) @_sigil_name
   (quoted_content) @injection.content)
- (#any-of? @_sigil_name "H" "LVN" "HOLO")
+ (#eq? @_sigil_name "HOLO")
  (#set! injection.language "heex")
- (#set! injection.combined))
-
-; Regex
-((sigil
-  (sigil_name) @_sigil_name
-  (quoted_content) @injection.content)
- (#any-of? @_sigil_name "r" "R")
- (#set! injection.language "regex")
- (#set! injection.combined))
-
-; Json
-((sigil
-  (sigil_name) @_sigil_name
-  (quoted_content) @injection.content)
- (#any-of? @_sigil_name "j" "J")
- (#set! injection.language "json")
  (#set! injection.combined))
 
 ; SQL injection
@@ -90,28 +74,12 @@
  (#set! injection.language "sql")
  (#set! injection.combined))
 
-; Surface
-((sigil
-  (sigil_name) @_sigil_name
-  (quoted_content) @injection.content)
- (#eq? @_sigil_name "F")
- (#set! injection.language "surface")
- (#set! injection.combined))
-
 ; Markdown
 ((sigil
   (sigil_name) @_sigil_name
   (quoted_content) @injection.content)
  (#eq? @_sigil_name "MD")
  (#set! injection.language "markdown")
- (#set! injection.combined))
-
-; Zig
-((sigil
-  (sigil_name) @_sigil_name
-  (quoted_content) @injection.content)
- (#any-of? @_sigil_name "z" "Z")
- (#set! injection.language "zig")
  (#set! injection.combined))
 
 ; Python