+++ crates/lumis-cli/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## [0.1.3](https://github.com/leandrocp/lumis/compare/rust-cli@v0.1.2...rust-cli@v0.1.3) (2026-04-04)
+
+
+### Bug Fixes
+
+* stop pinning lumis-core workspace deps ([ef2f002](https://github.com/leandrocp/lumis/commit/ef2f002eb60cd408872c5ce5f4dd7792bb00c557))
+
+## [0.1.2](https://github.com/leandrocp/lumis/compare/rust-cli@v0.1.1...rust-cli@v0.1.2) (2026-04-03)
+
+
+### Bug Fixes
+
+* loosen lumis-core dep ([61a0e45](https://github.com/leandrocp/lumis/commit/61a0e45bc877f3a2f4652660e8c38a282b4d4552))
+* query overrides and injections ([#551](https://github.com/leandrocp/lumis/issues/551)) ([2a356fb](https://github.com/leandrocp/lumis/commit/2a356fbd8f0841c7e0fdb5835330f98621c199ce))
+
+
 ## [0.1.1](https://github.com/leandrocp/lumis/compare/rust-cli@v0.1.0...rust-cli@v0.1.1) (2026-04-02)