+++ packages/javascript/wasm-bundle-backend/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.0.3](https://github.com/leandrocp/lumis/compare/wasm-bundle-backend@v0.0.2...wasm-bundle-backend@v0.0.3) (2026-04-05)
+
+
+### Features
+
+* language bundles ([#556](https://github.com/leandrocp/lumis/issues/556)) ([6e98734](https://github.com/leandrocp/lumis/commit/6e98734c20d8a86384f3e8b8e2b694c40bbb0f12))
+
 ## [0.0.2](https://github.com/leandrocp/lumis/compare/wasm-bundle-backend@v0.0.1...wasm-bundle-backend@v0.0.2) (2026-03-31)